% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{interpolacao}
\alias{interpolacao}
\title{estimativa de valores atraves da interpolacao}
\usage{
interpolacao(x1, x2, xk, y1, y2)
}
\description{
x1 primeiro valor de x
}
\details{
x2 segundo valor de x

xk valor que se quer descobrir o correspondente yk

y1 primeiro valor de y

y2 segundo valor de y
}
